% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaborMarketClass.R
\docType{class}
\name{LaborMarket-class}
\alias{LaborMarket-class}
\title{S4 Class for Representing a Labor Market}
\description{
A class to represent the state of a labor market including parameters and data.
}
\section{Slots}{

\describe{
\item{\code{panel}}{A data.table object representing panel data.}

\item{\code{init.params}}{A list of initial parameters.}

\item{\code{tranM}}{An array representing transition matrices.}

\item{\code{steadyM}}{An array representing steady-state matrices.}

\item{\code{alpha_mean}}{Mean of the alpha parameter.}

\item{\code{psi_mean}}{Mean of the psi parameter.}

\item{\code{psi_sd}}{Standard deviation of the psi parameter.}

\item{\code{alpha_sd}}{Standard deviation of the alpha parameter.}

\item{\code{csort}}{Cost of sorting.}

\item{\code{cnetw}}{Cost of networking.}

\item{\code{csig}}{Cost of signalling.}

\item{\code{fsize}}{Firm size.}

\item{\code{w_sigma}}{Standard deviation of wages.}

\item{\code{neduc}}{Number of education levels.}

\item{\code{sort_gap}}{Sorting gap.}

\item{\code{shocks}}{Shocks parameter.}
}}

