% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmbias.R
\name{lmbias}
\alias{lmbias}
\title{Corrects the Limited Mobility Bias Using Bootstrap}
\usage{
lmbias(
  femodel,
  data,
  vcov = "HC1",
  R = 1000,
  parallel = c("no", "multicore", "snow"),
  cluster = NULL,
  ncpus = getOption("boot.ncpus", 1L),
  cl = NULL
)
}
\arguments{
\item{femodel}{A model object of class \code{fixest} (from the \code{lfe} package) that contains the
fixed effects estimates to be corrected.}

\item{data}{A data.table object that contains the dataset used for the fixed effects model.}

\item{vcov}{A character string specifying the type of heteroskedasticity correction to be
used. Currently, only "HC0" and "HC1" are supported.}

\item{R}{An integer that sets the number of bootstrap replications. Default is 1000.}

\item{parallel}{A character string that determines the type of parallel operation to be used
(if any). Possible values are "no", "multicore", and "snow". Default is "no".}

\item{cluster}{An optional vector or factor that provides the data's cluster IDs if cluster
bootstrapping is desired.}

\item{ncpus}{An integer specifying the number of CPUs to be used for parallel processing.
This argument is relevant only if \code{parallel} is set to "multicore" or "snow".
Default uses the option set in \code{getOption("boot.ncpus", 1L)}.}

\item{cl}{An optional parallel or snow cluster for use with \code{parallel = "snow"}. If not
supplied and \code{parallel = "snow"}, a cluster on the local machine is created.}
}
\value{
A matrix that presents the variance, covariance, and correlation for the original
fixed effects, the estimated bias, and the corrected measures.
}
\description{
The \code{lmbias} function applies a bootstrapping technique to correct the limited mobility bias
(often referred to as small sample bias) in fixed-effects models. This implementation is based
on the method proposed by Azkarate-Askasua and Zerecero (2022). It uses a Rademacher random
vector and the fixest package's model specifications for its calculations.
}
\examples{
# Assuming example data and model exist:
# data <- data.table(...)
# model <- feols(y ~ x1 + x2 | fe1 + fe2, data = data)
# lmbias(model, data)
}
